+++ projects/docs/src/app/i18n/types/seo.ts
@@ -16,6 +16,8 @@ export interface DocsSeoMessages {
   guidesSignalForms: DocsPageSeo;
   guidesApiConventions: DocsPageSeo;
   guidesFloatingUi: DocsPageSeo;
+  guidesComposition: DocsPageSeo;
+  guidesRecipes: DocsPageSeo;
   guidesCrudDemo: DocsPageSeo;
   maturity: DocsPageSeo;
   designTokens: DocsPageSeo;